Forgot password using otp

+++ src/modules/user/user.service.ts
@@ -1,25 +1,26 @@
 import { ConflictException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
+import { hash } from 'bcrypt';
 import { User } from 'src/models';
 
 @Injectable()
 export class UserService {
   constructor(@InjectModel(User) private userModel: typeof User) {}
 
-  async findUserById(id: number) {
-    return this.userModel.findOne({
-      where: { id },
+  async findUserById(userId: number) {
+    return await this.userModel.findOne({
+      where: { id: userId },
       attributes: { exclude: ['password', 'refreshToken'] },
       raw: true,
     });
   }
 
   async findUserByEmail(email: string) {
-    return this.userModel.findOne({ where: { email } });
+    return await this.userModel.findOne({ where: { email } });
   }
 
   async findUserByPhoneNumber(phoneNumber: string) {
-    return this.userModel.findOne({ where: { phoneNumber } });
+    return await this.userModel.findOne({ where: { phoneNumber } });
   }
 
   async createUser(
@@ -28,24 +29,59 @@ export class UserService {
     name: string,
     phoneNumber: string,
   ) {
-    return this.userModel.create({ email, password, name, phoneNumber });
+    const hashedPassword = await hash(password, 10);
+
+    return await this.userModel.create({
+      email,
+      password: hashedPassword,
+      name,
+      phoneNumber,
+    });
+    ``;
   }
 
-  async updateUser(id: number, name: string, phoneNumber: string) {
-    return this.userModel.update({ name, phoneNumber }, { where: { id } });
+  async updateUser(
+    userId: number,
+    newUserData: {
+      email?: string;
+      password?: string;
+      name?: string;
+      phoneNumber?: string;
+      isVerified?: boolean;
+      refreshToken?: string;
+    },
+  ) {
+    const updatedUser = await this.userModel.update(
+      { ...newUserData },
+      { where: { id: userId }, returning: true },
+    );
+
+    return updatedUser[1];
+  }
+
+  async updateUserRefreshToken(userId: number, refreshToken: string) {
+    return await this.updateUser(userId, { refreshToken });
   }
 
   async updateEmailVerificationStatus(userId: number, isVerified: boolean) {
     if (isVerified) throw new ConflictException('User is already verified.');
 
-    const updatedUser = await this.userModel.update(
-      { isVerified },
-      { where: { id: userId }, returning: true },
-    );
+    const updatedUser = await this.updateUser(userId, { isVerified });
 
     return updatedUser;
   }
-  async deleteUser(id: number) {
-    return this.userModel.destroy({ where: { id } });
+
+  async updatePassword(userId: number, newPassword: string) {
+    const newHashedPassword = await hash(newPassword, 10);
+
+    const updatedUser = await this.updateUser(userId, {
+      password: newHashedPassword,
+    });
+
+    return updatedUser;
+  }
+
+  async deleteUser(userId: number) {
+    return await this.userModel.destroy({ where: { id: userId } });
   }
 }
